split search provider into separate queries hooks
